Octave 7 compatibility fix: rmdir() now errors out if the directory is nonexistent
(cherry picked from commit 418c290eada40e971e2c4ab176a629efbd47d58f)

+++ tests/run_block_byte_tests_octave.m
@@ -61,8 +61,10 @@ for blockFlag = 0:1
 
         # Workaround for strange race condition related to the static/dynamic
         # files (especially when we switch to/from use_dll)
-        rmdir('+ls2003_tmp', 's');
-        pause(1)
+        if exist('+ls2003_tmp')
+            rmdir('+ls2003_tmp', 's');
+            pause(1)
+        endif
 
         for i = 1:length(solve_algos)
             num_block_tests = num_block_tests + 1;
